Maybe its ready
I just add a new if in the function "calculateTotalPrice" ¿Why? Ok because when one checkbox change "True/False" trigger de if, if a true the variable "selectBox" take the value (Variable of reference) of the variable "checkboxes". Now in the "for" is execute a new if.  If the name of the parts in the list == the id of the checkboxes so theTotalPriceOfParts = The price of the parts in the list. While all this is true the totalPriceOfParts+=selectCarPrice.

Uf I hope you stand all this xd. Here i dont use the translate.